# openapi.yaml
openapi: 3.0.3
info:
  title: Oh Your Ear API
  description: API for the Oh Your Ear ear training application
  version: 1.0.0
servers:
  - url: /api/v1
    description: Local development
paths:
...
  /me/avatar:
    get:
      operationId: getMyAvatar
      summary: Get the avatar this user uploaded
      security:
        - bearerAuth: []
      responses:
        '200':
          description: The uploaded picture
          content:
            image/png:
              schema:
                type: string
                format: binary
            image/jpeg:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: This user has not uploaded one
    post:
      operationId: putMyAvatar
      summary: Upload or replace this user's avatar
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required: [file]
      responses:
        '200':
          description: The user, with the new avatar URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '413':
          description: The picture is larger than the limit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      operationId: deleteMyAvatar
      summary: Remove this user's uploaded avatar
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Removed; the account falls back to its provider picture or none
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserResponse:
      type: object
      required:
        - id
        - email
        - hasPassword
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        hasPassword:
          type: boolean
          description: >
            Whether this account can also sign in with a password. The client uses it to
            offer setting one, and to choose which field to show first.
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
        message:
          type: string
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'